fix to number controllers text fields on .listen()
Stop gap fix I used on my app to allow edits to the text inputs on number controllers in the instance where .listen() has been used. A true fix would probably temporarily remove a controller from the listened object array, or set a flag on the record in the array, once the text input is given focus/on the first keypress/input event. Once a focus cancelling action is caught, the listening could be reinstated. 

If you somehow tracked when both changes to the object's variable that you're looking at and the input field have happened, you could have a dialog pop up asking the user which to use....although, the object variable could keep changing while this pop up is displayed, so it's probably not worth it.